requirements.txt
pyodbc
pandas
python-dotenv
openpyxl